+++ CHANGES.rst
@@ -183,6 +183,9 @@ astropy.table
 - Added support for stacking ``Column``, mixin column (e.g. ``Quantity``,
   ``Time``) or column-like objects. [#7674]
 
+- Added support for inserting a row into a Table that has ``Time`` or
+  ``TimeDelta`` column(s). [#7897]
+
 astropy.tests
 ^^^^^^^^^^^^^
 
@@ -211,6 +214,8 @@ astropy.time
 - Add fractional second support for ``strftime`` and ``strptime`` methods
   of ``Time`` class. [#7705]
 
+- Added an ``insert`` method to allow inserting one or more values into a
+  ``Time`` or ``TimeDelta`` object. [#7897]
 
 astropy.uncertainty
 ^^^^^^^^^^^^^^^^^^^